refactor(chat): extract artifact file actions
